добавлен опциональный обработчик (resolve_model_reference): он срабатывает только когда model в кортеже — строка (новый формат store_references=True). Если там объект — сразу возвращает его, и старый формат работает без изменений. one_model_predict всегда получает уже готовый объект модели — ему всё равно, откуда он пришёл.
